[DDW-378] remove any reference to node modules in pure frontend code
