+++ source/renderer/app/components/sidebar/wallets/SidebarWalletsMenu.js
@@ -7,7 +7,7 @@ import SidebarSubMenu from '../SidebarMenu';
 import styles from './SidebarWalletsMenu.scss';
 import addWalletIcon from '../../../assets/images/sidebar/add-wallet-ic.inline.svg';
 import SidebarWalletMenuItem from './SidebarWalletMenuItem';
-import type { SidebarWalletType } from '../../../stores/SidebarStore';
+import type { SidebarWalletType } from '../../../types/sidebarTypes';
 
 const messages = defineMessages({
   addAdaWallet: {
