+++ source/renderer/app/stores/SidebarStore.js
@@ -1,5 +1,5 @@
 // @flow
-import { observable, action, computed } from 'mobx';
+import { action, computed, observable } from 'mobx';
 import { get } from 'lodash';
 import { ipcRenderer } from 'electron';
 import Store from './lib/Store';
@@ -7,6 +7,7 @@ import resolver from '../utils/imports';
 import environment from '../../../common/environment';
 import { syncStateTags } from '../domains/Wallet';
 import { GO_TO_ADA_REDEMPTION_SCREEN_CHANNEL } from '../../../common/ipc-api/go-to-ada-redemption-screen';
+import type { SidebarWalletType } from '../types/sidebarTypes';
 
 const sidebarConfig = resolver('config/sidebarConfig');
 const { formattedWalletAmount } = resolver('utils/formatters');
@@ -97,12 +98,3 @@ export default class SidebarStore extends Store {
   };
 
 }
-
-export type SidebarWalletType = {
-  id: string,
-  title: string,
-  info: string,
-  isConnected: bool,
-  isRestoreActive: bool,
-  restoreProgress: number,
-};
