+++ source/renderer/app/types/sidebarTypes.js
@@ -1,0 +1,8 @@
+export type SidebarWalletType = {
+  id: string,
+  title: string,
+  info: string,
+  isConnected: bool,
+  isRestoreActive: bool,
+  restoreProgress: number,
+};
